refactor(integrations): split booking/mod.rs OAuth + retry into submodules
Churn-hotspot reduction on backend/crates/integrations/src/booking/mod.rs
(3185 lines). Pure verbatim move — no behaviour change:

- extract the OAuth 2.0 authorization-code client (BookingOAuthConfig /
  BookingOAuthTokens / BookingOAuthClient / OAuthTokenResponse + consts +
  its oauth_tests) into booking/oauth.rs
- extract the AC-5 retry policy + push-outcome types (BookingRetryConfig,
  PushOutcome, is_retryable_status) into booking/retry.rs

Public paths preserved via `pub use` re-exports in mod.rs, so
`integrations::booking::{...}` and the crate-level `pub use booking::{...}`
in lib.rs keep resolving unchanged. Sole visibility change:
is_retryable_status is now `pub(super)` (module-boundary requirement),
re-imported into mod.rs so callers keep using it unqualified. Intra-doc
links to BookingClient qualified with `super::` from the moved files.

mod.rs: 3185 -> 2786 lines.
