+++ backend/crates/integrations/src/booking/mod.rs
@@ -29,6 +29,27 @@ pub const OTA_NAMESPACE: &str = "http://www.opentravel.org/OTA/2003/05";
 /// Booking.com OTA wire format. See [`ota_xml`] for details.
 pub mod ota_xml;
 
+/// OAuth 2.0 (authorization-code) client for Booking.com's newer Connectivity
+/// APIs. See [`oauth`] for details.
+pub mod oauth;
+
+/// Retry policy + push-outcome types for OTA rate/availability push (AC-5).
+/// See [`retry`] for details.
+pub mod retry;
+
+// Re-export the moved public items so existing `booking::<Item>` paths (and the
+// crate-level `pub use booking::{...}` in `lib.rs`) keep resolving unchanged
+// after the module split.
+pub use oauth::{
+    BookingOAuthClient, BookingOAuthConfig, BookingOAuthTokens, BOOKING_OAUTH_AUTH_URL,
+    BOOKING_OAUTH_TOKEN_URL,
+};
+pub use retry::{BookingRetryConfig, PushOutcome};
+// `is_retryable_status` is crate-private (pub(super)); bring it into module
+// scope so the `BookingClient` impl and the in-module tests keep calling it
+// unqualified.
+use retry::is_retryable_status;
+
 // ============================================
 // Error Types
 // ============================================
@@ -929,154 +950,6 @@ pub struct RoomTypeMapping {
 }
 
 // ============================================
-// Retry Configuration (AC-5 rate/availability push)
-// ============================================
-
-/// Retry policy for OTA rate & availability *push* operations.
-///
-/// Booking.com's Supply XML endpoint is occasionally transient-fail under
-/// load (HTTP 429/5xx) and the OTA push messages are idempotent (each carries
-/// an absolute `Start`/`End`/`InvTypeCode` application control rather than a
-/// delta), so re-sending the same document is safe. This struct controls the
-/// bounded exponential-backoff retry loop used by [`BookingClient::push_rates`]
-/// and [`BookingClient::push_availability`].
-///
-/// This is intentionally a small, push-specific policy distinct from
-/// [`crate::connector::RetryConfig`], which drives the generic [`crate::connector::Connector`]
-/// request/response abstraction. The Booking client talks to `reqwest`
-/// directly (it must hand-roll OTA XML bodies + Basic auth), so it carries its
-/// own lightweight policy rather than routing through the generic connector.
-#[derive(Debug, Clone)]
-pub struct BookingRetryConfig {
-    /// Total number of attempts (1 = no retry). Must be >= 1.
-    pub max_attempts: u32,
-    /// Delay before the first retry, in milliseconds.
-    pub initial_delay_ms: u64,
-    /// Upper bound on any single backoff delay, in milliseconds.
-    pub max_delay_ms: u64,
-    /// Multiplier applied to the delay after each failed attempt.
-    pub backoff_multiplier: u32,
-}
-
-impl Default for BookingRetryConfig {
-    fn default() -> Self {
-        Self {
-            max_attempts: 3,
-            initial_delay_ms: 500,
-            max_delay_ms: 8_000,
-            backoff_multiplier: 2,
-        }
-    }
-}
-
-impl BookingRetryConfig {
-    /// A no-retry policy (single attempt) — useful in tests to keep them fast.
-    pub fn no_retry() -> Self {
-        Self {
-            max_attempts: 1,
-            initial_delay_ms: 0,
-            max_delay_ms: 0,
-            backoff_multiplier: 1,
-        }
-    }
-
-    /// Return a copy of this policy with `max_attempts` overridden, keeping the
-    /// other (delay/backoff) parameters. Handy for zero-delay multi-attempt
-    /// retry configs in tests (`no_retry().clone_with_attempts(3)`).
-    pub fn clone_with_attempts(&self, max_attempts: u32) -> Self {
-        Self {
-            max_attempts: max_attempts.max(1),
-            ..self.clone()
-        }
-    }
-
-    /// Backoff delay (ms) before the retry that follows `attempt` (0-based:
-    /// `attempt = 0` is the delay after the first attempt failed). Capped at
-    /// [`Self::max_delay_ms`].
-    pub fn delay_for_attempt(&self, attempt: u32) -> u64 {
-        let factor = self.backoff_multiplier.saturating_pow(attempt);
-        self.initial_delay_ms
-            .saturating_mul(factor as u64)
-            .min(self.max_delay_ms)
-    }
-
-    /// Effective delay (ms) to wait before the retry that follows the failed
-    /// attempt indexed by `prev_attempt` (0-based, same convention as
-    /// [`Self::delay_for_attempt`]), combining our exponential backoff with an
-    /// optional server-supplied `Retry-After` hint.
-    ///
-    /// `retry_after_ms` is the hint already converted to milliseconds and
-    /// capped by the caller. The larger of the two values wins: we never retry
-    /// sooner than the server explicitly asked for (`Retry-After`), and never
-    /// sooner than our own backoff schedule would allow. When no hint is
-    /// present this is exactly [`Self::delay_for_attempt`].
-    ///
-    /// Extracted from the inline logic in
-    /// [`BookingClient::post_ota_with_retry`] so the delay computation is
-    /// unit-testable without spinning up a mock server and real timers.
-    pub fn effective_delay_ms(&self, prev_attempt: u32, retry_after_ms: Option<u64>) -> u64 {
-        let base = self.delay_for_attempt(prev_attempt);
-        match retry_after_ms {
-            Some(hint) => hint.max(base),
-            None => base,
-        }
-    }
-}
-
-/// Outcome of a combined availability + rate *push* sync to Booking.com
-/// (AC-5).
-///
-/// The two OTA streams — `OTA_HotelAvailNotifRQ` (availability) and
-/// `OTA_HotelRateAmountNotifRQ` (rates) — are pushed independently and can
-/// succeed or fail independently. Collapsing them into a single
-/// `Result<(), _>` loses that distinction: a caller could not tell whether an
-/// error meant "nothing was applied" or "availability landed but rates were
-/// rejected", leaving the channel in a half-synced state with no signal.
-///
-/// [`BookingClient::push_availability_and_rates`] returns this struct so the
-/// caller sees exactly which stream applied, how many messages each carried,
-/// and the per-stream error text when one failed. Each stream still runs
-/// through [`BookingClient::post_ota_with_retry`], so transient failures are
-/// retried with the same bounded backoff + idempotency-token semantics as the
-/// single-stream pushes.
-#[derive(Debug, Clone, Default, PartialEq, Eq)]
-pub struct PushOutcome {
-    /// Number of availability messages sent (0 when none were supplied).
-    pub availability_pushed: usize,
-    /// Number of rate messages sent (0 when none were supplied).
-    pub rates_pushed: usize,
-    /// `Some(error)` if the availability push was attempted and failed;
-    /// `None` if it succeeded or was skipped (no availability updates).
-    pub availability_error: Option<String>,
-    /// `Some(error)` if the rate push was attempted and failed; `None` if it
-    /// succeeded or was skipped (no rate updates).
-    pub rates_error: Option<String>,
-}
-
-impl PushOutcome {
-    /// True when neither stream reported an error (a fully-applied or no-op
-    /// sync).
-    pub fn is_success(&self) -> bool {
-        self.availability_error.is_none() && self.rates_error.is_none()
-    }
-
-    /// True when exactly one of the two streams failed while the other
-    /// applied — the channel is now in a half-synced state the caller must
-    /// reconcile.
-    pub fn is_partial(&self) -> bool {
-        self.availability_error.is_some() != self.rates_error.is_some()
-    }
-}
-
-/// Whether an HTTP status code returned by the Booking.com endpoint is worth
-/// retrying. Transient server-side / throttling failures are retryable;
-/// client errors (4xx other than 408/429) are not — they indicate a bad
-/// message and will fail identically on retry.
-fn is_retryable_status(status: u16) -> bool {
-    matches!(status, 408 | 425 | 429 | 500 | 502 | 503 | 504)
-}
-
-// ============================================
 // Client Implementation
 // ============================================
 
@@ -1613,280 +1486,8 @@ impl BookingClient {
 }
 
 // ============================================
-// OAuth 2.0 (authorization-code) client
-// ============================================
-//
-// Booking.com's newer Connectivity APIs use an OAuth 2.0 authorization-code
-// grant in addition to the legacy Basic-auth Supply XML credentials handled by
-// [`BookingClient`].  This section mirrors the Airbnb OAuth client
-// (`integrations::airbnb`) so the api-server can host a parallel
-// `POST /organizations/{org_id}/booking/token/exchange` route (Coverage 83-2).
-
-/// Booking.com OAuth token endpoint (authorization-code + refresh grants).
-pub const BOOKING_OAUTH_TOKEN_URL: &str = "https://account.booking.com/oauth2/token";
-
-/// Booking.com OAuth authorization endpoint (front-channel redirect target).
-pub const BOOKING_OAUTH_AUTH_URL: &str = "https://account.booking.com/oauth2/authorize";
-
-/// OAuth configuration for Booking.com.
-#[derive(Debug, Clone)]
-pub struct BookingOAuthConfig {
-    /// Booking.com OAuth client ID.
-    pub client_id: String,
-    /// Booking.com OAuth client secret.
-    pub client_secret: String,
-    /// OAuth redirect URI for the callback.
-    pub redirect_uri: String,
-}
-
-/// OAuth tokens received from Booking.com.
-#[derive(Debug, Clone, Serialize, Deserialize)]
-pub struct BookingOAuthTokens {
-    /// Access token for API calls.
-    pub access_token: String,
-    /// Refresh token for obtaining new access tokens.
-    pub refresh_token: Option<String>,
-    /// When the access token expires.
-    pub expires_at: Option<DateTime<Utc>>,
-    /// Token type (usually "Bearer").
-    pub token_type: String,
-    /// OAuth scopes granted.
-    pub scope: Option<String>,
-}
-
-/// Booking.com OAuth client (authorization-code grant).
-///
-/// Distinct from [`BookingClient`], which speaks the legacy Basic-auth Supply
-/// XML protocol.  This client only handles the OAuth token lifecycle.
-pub struct BookingOAuthClient {
-    client: reqwest::Client,
-    config: BookingOAuthConfig,
-}
-
-impl BookingOAuthClient {
-    /// Create a new Booking.com OAuth client.
-    pub fn new(config: BookingOAuthConfig) -> Self {
-        Self {
-            client: reqwest::Client::new(),
-            config,
-        }
-    }
-
-    /// Create a new client from individual OAuth parameters.
-    pub fn with_credentials(
-        client_id: String,
-        client_secret: String,
-        redirect_uri: String,
-    ) -> Self {
-        Self::new(BookingOAuthConfig {
-            client_id,
-            client_secret,
-            redirect_uri,
-        })
-    }
-
-    /// Generate the OAuth authorization URL the user is redirected to.
-    pub fn generate_auth_url(&self, state: &str) -> String {
-        let scopes = "reservations:read availability:write rates:write";
-        format!(
-            "{}?client_id={}&redirect_uri={}&response_type=code&state={}&scope={}",
-            BOOKING_OAUTH_AUTH_URL,
-            urlencoding::encode(&self.config.client_id),
-            urlencoding::encode(&self.config.redirect_uri),
-            urlencoding::encode(state),
-            urlencoding::encode(scopes),
-        )
-    }
-
-    /// Exchange an authorization code for access and refresh tokens.
-    pub async fn exchange_code(&self, code: &str) -> Result<BookingOAuthTokens, BookingError> {
-        #[derive(Serialize)]
-        struct TokenRequest<'a> {
-            grant_type: &'a str,
-            client_id: &'a str,
-            client_secret: &'a str,
-            code: &'a str,
-            redirect_uri: &'a str,
-        }
-
-        let response = self
-            .client
-            .post(BOOKING_OAUTH_TOKEN_URL)
-            .form(&TokenRequest {
-                grant_type: "authorization_code",
-                client_id: &self.config.client_id,
-                client_secret: &self.config.client_secret,
-                code,
-                redirect_uri: &self.config.redirect_uri,
-            })
-            .send()
-            .await?;
-
-        if !response.status().is_success() {
-            let error = response.text().await.unwrap_or_default();
-            return Err(BookingError::Auth(error));
-        }
-
-        let token_response: OAuthTokenResponse = response
-            .json()
-            .await
-            .map_err(|e| BookingError::Api(e.to_string()))?;
-
-        Ok(token_response.into_tokens(None))
-    }
-
-    /// Refresh an expired access token using the stored refresh token.
-    pub async fn refresh_token(
-        &self,
-        refresh_token: &str,
-    ) -> Result<BookingOAuthTokens, BookingError> {
-        #[derive(Serialize)]
-        struct RefreshRequest<'a> {
-            grant_type: &'a str,
-            client_id: &'a str,
-            client_secret: &'a str,
-            refresh_token: &'a str,
-        }
-
-        let response = self
-            .client
-            .post(BOOKING_OAUTH_TOKEN_URL)
-            .form(&RefreshRequest {
-                grant_type: "refresh_token",
-                client_id: &self.config.client_id,
-                client_secret: &self.config.client_secret,
-                refresh_token,
-            })
-            .send()
-            .await?;
-
-        if !response.status().is_success() {
-            let error = response.text().await.unwrap_or_default();
-            return Err(BookingError::Auth(error));
-        }
-
-        let token_response: OAuthTokenResponse = response
-            .json()
-            .await
-            .map_err(|e| BookingError::Api(e.to_string()))?;
-
-        Ok(token_response.into_tokens(Some(refresh_token)))
-    }
-}
-
-/// Raw token-endpoint response shape (shared by exchange + refresh).
-#[derive(Debug, Deserialize)]
-struct OAuthTokenResponse {
-    access_token: String,
-    refresh_token: Option<String>,
-    expires_in: Option<i64>,
-    #[serde(default = "default_token_type")]
-    token_type: String,
-    scope: Option<String>,
-}
-
-fn default_token_type() -> String {
-    "Bearer".to_string()
-}
-
-impl OAuthTokenResponse {
-    /// Convert the wire response into [`BookingOAuthTokens`], computing
-    /// `expires_at` from `expires_in` and falling back to `fallback_refresh`
-    /// when the endpoint did not rotate the refresh token.
-    fn into_tokens(self, fallback_refresh: Option<&str>) -> BookingOAuthTokens {
-        let expires_at = self
-            .expires_in
-            .map(|secs| Utc::now() + Duration::seconds(secs));
-        BookingOAuthTokens {
-            access_token: self.access_token,
-            refresh_token: self
-                .refresh_token
-                .or_else(|| fallback_refresh.map(|s| s.to_string())),
-            expires_at,
-            token_type: self.token_type,
-            scope: self.scope,
-        }
-    }
-}
-
-// ============================================
 // Tests
 // ============================================
-
-#[cfg(test)]
-mod oauth_tests {
-    use super::*;
-
-    #[test]
-    fn test_generate_auth_url() {
-        let client = BookingOAuthClient::new(BookingOAuthConfig {
-            client_id: "test_client_id".to_string(),
-            client_secret: "test_secret".to_string(),
-            redirect_uri: "http://localhost:8080/callback".to_string(),
-        });
-        let url = client.generate_auth_url("test_state");
-        assert!(url.contains("account.booking.com"));
-        assert!(url.contains("test_client_id"));
-        assert!(url.contains("test_state"));
-        assert!(url.contains("response_type=code"));
-        assert!(!url.contains("test_secret"));
-    }
-
-    #[test]
-    fn test_token_response_deserialization_full() {
-        let json = r#"{"access_token":"at-123","refresh_token":"rt-456","expires_in":3600,"token_type":"Bearer","scope":"reservations:read"}"#;
-        let parsed: OAuthTokenResponse = serde_json::from_str(json).unwrap();
-        let tokens = parsed.into_tokens(None);
-        assert_eq!(tokens.access_token, "at-123");
-        assert_eq!(tokens.refresh_token.as_deref(), Some("rt-456"));
-        assert_eq!(tokens.token_type, "Bearer");
-        assert_eq!(tokens.scope.as_deref(), Some("reservations:read"));
-        assert!(tokens.expires_at.is_some());
-    }
-
-    #[test]
-    fn test_token_response_defaults_token_type() {
-        let json = r#"{"access_token":"at-789","expires_in":7200}"#;
-        let parsed: OAuthTokenResponse = serde_json::from_str(json).unwrap();
-        let tokens = parsed.into_tokens(None);
-        assert_eq!(tokens.token_type, "Bearer");
-        assert!(tokens.refresh_token.is_none());
-        assert!(tokens.scope.is_none());
-    }
-
-    #[test]
-    fn test_refresh_falls_back_to_existing_refresh_token() {
-        let json = r#"{"access_token":"at-new","expires_in":3600,"token_type":"Bearer"}"#;
-        let parsed: OAuthTokenResponse = serde_json::from_str(json).unwrap();
-        let tokens = parsed.into_tokens(Some("rt-original"));
-        assert_eq!(tokens.access_token, "at-new");
-        assert_eq!(tokens.refresh_token.as_deref(), Some("rt-original"));
-    }
-
-    #[test]
-    fn test_rotated_refresh_token_wins_over_fallback() {
-        let json = r#"{"access_token":"at-new","refresh_token":"rt-rotated","expires_in":3600}"#;
-        let parsed: OAuthTokenResponse = serde_json::from_str(json).unwrap();
-        let tokens = parsed.into_tokens(Some("rt-original"));
-        assert_eq!(tokens.refresh_token.as_deref(), Some("rt-rotated"));
-    }
-
-    #[test]
-    fn test_oauth_tokens_roundtrip_serde() {
-        let tokens = BookingOAuthTokens {
-            access_token: "at".to_string(),
-            refresh_token: Some("rt".to_string()),
-            expires_at: None,
-            token_type: "Bearer".to_string(),
-            scope: Some("rates:write".to_string()),
-        };
-        let json = serde_json::to_string(&tokens).unwrap();
-        let back: BookingOAuthTokens = serde_json::from_str(&json).unwrap();
-        assert_eq!(back.access_token, "at");
-        assert_eq!(back.refresh_token.as_deref(), Some("rt"));
-        assert_eq!(back.scope.as_deref(), Some("rates:write"));
-    }
-}
 
 #[cfg(test)]
 mod tests {
